Fix index type so that it compare with size()
1

+++ src/collection.cpp
@@ -79,7 +79,7 @@ namespace seneca {
 	};
 
 	ostream& operator<<(ostream& os, const Collection& collection) {
-		for (int i = 0; i < collection.size(); i++) {
+		for (size_t i = 0; i < collection.size(); i++) {
 			collection[i]->display(os);
 		};
