+++ src/spellChecker.cpp
@@ -56,7 +56,7 @@ namespace seneca {
 	void SpellChecker::showStatistics(std::ostream& out) const {
 		out <<	"Spellchecker Statistics" << endl;
 
-		for(int i = 0; i < SpellChecker::counter.size(); i++ ) {
+		for(size_t i = 0; i < SpellChecker::counter.size(); i++ ) {
 			out << left << setw(15) << m_badWords[i] << ": " << SpellChecker::counter[i] << " replacements" << endl;
 		};
 	};
